+++ lib/modules/storage/storage_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: storage
-// Updated at: 2026-05-23 22:21:49
-// Commit: feat(storage): add async storage stream provider
+// Updated at: 2026-05-24 09:28:26
+// Commit: fix(storage): fix memory leak by cancelling storage subscription
 
 class storageComponentState {
-  final String commitRef = '1370';
-  final String lastMessage = 'feat(storage): add async storage stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-23T22:21:49+05:00');
+  final String commitRef = '1374';
+  final String lastMessage = 'fix(storage): fix memory leak by cancelling storage subscription';
+  final DateTime timestamp = DateTime.parse('2026-05-24T09:28:26+05:00');
 }